backend/requirements.txt
fastapi[all]
uvicorn
pandas
numpy
scikit-learn
requests
python-dotenv